refactor: create folder for each type of algorithms
