--- algorithms/sorting.js
@@ -1,435 +1,0 @@
-/**
- * Bubble Sort generator.
- * Yields objects of the form:
- *   { type: 'compare', indices: [i, j] }
- *   { type: 'swap', indices: [i, j], array: [current array state] }
- * @param {number[]} arr - Array to sort
- * @yields {{type: string, indices: number[], array?: number[]}}
- */
-export function* bubbleSort(arr) {
-  const steps = [...arr];
-  for (let i = 0; i < steps.length; i++) {
-    for (let j = 0; j < steps.length - i - 1; j++) {
-      yield { type: 'compare', indices: [j, j + 1] };
-      if (steps[j] > steps[j + 1]) {
-        [steps[j], steps[j + 1]] = [steps[j + 1], steps[j]];
-        yield { type: 'swap', indices: [j, j + 1], array: [...steps] };
-      }
-    }
-  }
-}
-
-/**
- * Selection Sort generator.
- * Yields objects of the form:
- *   { type: 'compare', indices: [current, min] }
- *   { type: 'newMin', indices: [min], minIndex: number }
- *   { type: 'swap', indices: [i, minIndex], array: [current array state] }
- *   { type: 'sorted', indices: [i] } (when position i is finalized)
- * @param {number[]} arr - Array to sort
- * @yields {{type: string, indices: number[], array?: number[], minIndex?: number}}
- */
-export function* selectionSort(arr) {
-  const steps = [...arr];
-
-  for (let i = 0; i < steps.length - 1; i++) {
-    let minIndex = i;
-
-    // Highlight the current position being filled
-    yield { type: 'sorting', indices: [i] };
-
-    // Find minimum element in remaining unsorted array
-    for (let j = i + 1; j < steps.length; j++) {
-      yield { type: 'compare', indices: [j, minIndex] };
-
-      if (steps[j] < steps[minIndex]) {
-        minIndex = j;
-        yield { type: 'newMin', indices: [minIndex], minIndex };
-      }
-    }
-
-    // Swap if minimum is not at current position
-    if (minIndex !== i) {
-      [steps[i], steps[minIndex]] = [steps[minIndex], steps[i]];
-      yield { type: 'swap', indices: [i, minIndex], array: [...steps] };
-    }
-
-    // Mark position as sorted
-    yield { type: 'sorted', indices: [i] };
-  }
-
-  // Mark the last element as sorted too
-  yield { type: 'sorted', indices: [steps.length - 1] };
-}
-
-/**
- * Insertion Sort generator.
- * Yields objects of the form:
- *   { type: 'current', indices: [current] } - Current element being inserted
- *   { type: 'compare', indices: [j, j+1] } - Comparing elements
- *   { type: 'shift', indices: [j], array: [current array state] } - Shifting element right
- *   { type: 'insert', indices: [position], array: [current array state] } - Inserting element
- *   { type: 'sorted', indices: [0, i] } - Elements 0 to i are sorted
- * @param {number[]} arr - Array to sort
- * @yields {{type: string, indices: number[], array?: number[]}}
- */
-export function* insertionSort(arr) {
-  const steps = [...arr];
-
-  for (let i = 1; i < steps.length; i++) {
-    const current = steps[i];
-    let j = i - 1;
-
-    // Highlight the current element being inserted
-    yield { type: 'current', indices: [i] };
-
-    // Move elements greater than current one position ahead
-    while (j >= 0 && steps[j] > current) {
-      // Show comparison
-      yield { type: 'compare', indices: [j, j + 1] };
-
-      // Shift element to the right
-      steps[j + 1] = steps[j];
-      yield { type: 'shift', indices: [j + 1], array: [...steps] };
-
-      j--;
-    }
-
-    // Insert the current element at its correct position
-    steps[j + 1] = current;
-    yield { type: 'insert', indices: [j + 1], array: [...steps] };
-
-    // Show that elements from 0 to i are now sorted
-    yield { type: 'sorted', indices: Array.from({ length: i + 1 }, (_, k) => k) };
-  }
-}
-
-/**
- * Merge Sort generator.
- * Yields objects of the form:
- *   { type: 'divide', indices: [left, right] } - Showing division of array
- *   { type: 'merge', indices: [left, mid, right] } - Starting merge operation
- *   { type: 'compare', indices: [i, j] } - Comparing elements during merge
- *   { type: 'place', indices: [position], array: [current array state] } - Placing element in merged position
- *   { type: 'copy', indices: [position], array: [current array state] } - Copying remaining elements
- *   { type: 'complete', indices: [left, right], array: [current array state] } - Merge complete for this range
- * @param {number[]} arr - Array to sort
- * @yields {{type: string, indices: number[], array?: number[]}}
- */
-export function* mergeSort(arr) {
-  const steps = [...arr];
-
-  function* mergeSortHelper(left, right) {
-    if (left >= right) return;
-
-    // Show division
-    yield { type: 'divide', indices: [left, right] };
-
-    const mid = Math.floor((left + right) / 2);
-
-    // Recursively sort left half
-    yield* mergeSortHelper(left, mid);
-
-    // Recursively sort right half
-    yield* mergeSortHelper(mid + 1, right);
-
-    // Merge the sorted halves
-    yield* merge(left, mid, right);
-  }
-
-  function* merge(left, mid, right) {
-    // Show merge operation starting
-    yield { type: 'merge', indices: [left, mid, right] };
-
-    // Create temporary arrays for left and right subarrays
-    const leftArr = [];
-    const rightArr = [];
-
-    // Copy data to temp arrays
-    for (let i = left; i <= mid; i++) {
-      leftArr.push(steps[i]);
-    }
-    for (let j = mid + 1; j <= right; j++) {
-      rightArr.push(steps[j]);
-    }
-
-    let i = 0; // Initial index of left subarray
-    let j = 0; // Initial index of right subarray
-    let k = left; // Initial index of merged subarray
-
-    // Merge the temp arrays back into steps[left..right]
-    while (i < leftArr.length && j < rightArr.length) {
-      // Show comparison
-      const leftOriginalIndex = left + i;
-      const rightOriginalIndex = mid + 1 + j;
-      yield { type: 'compare', indices: [leftOriginalIndex, rightOriginalIndex] };
-
-      if (leftArr[i] <= rightArr[j]) {
-        steps[k] = leftArr[i];
-        yield { type: 'place', indices: [k], array: [...steps] };
-        i++;
-      } else {
-        steps[k] = rightArr[j];
-        yield { type: 'place', indices: [k], array: [...steps] };
-        j++;
-      }
-      k++;
-    }
-
-    // Copy remaining elements of leftArr[], if any
-    while (i < leftArr.length) {
-      steps[k] = leftArr[i];
-      yield { type: 'copy', indices: [k], array: [...steps] };
-      i++;
-      k++;
-    }
-
-    // Copy remaining elements of rightArr[], if any
-    while (j < rightArr.length) {
-      steps[k] = rightArr[j];
-      yield { type: 'copy', indices: [k], array: [...steps] };
-      j++;
-      k++;
-    }
-
-    // Show completion of this merge
-    yield { type: 'complete', indices: [left, right], array: [...steps] };
-  }
-
-  // Start the merge sort process
-  yield* mergeSortHelper(0, steps.length - 1);
-}
-
-/**
- * Binary Search generator.
- * Requires a sorted array to work properly.
- * Yields objects of the form:
- *   { type: 'compare', indices: [mid], target: number, found: boolean }
- *   { type: 'range', indices: [left, right], target: number }
- *   { type: 'found', indices: [foundIndex], target: number } (if found)
- *   { type: 'not-found', target: number } (if not found)
- * @param {number[]} arr - Sorted array to search in
- * @param {number} target - Value to search for
- * @yields {{type: string, indices: number[], target: number, found?: boolean}}
- */
-export function* binarySearch(arr, target) {
-  let left = 0;
-  let right = arr.length - 1;
-
-  while (left <= right) {
-    // Show current search range
-    yield { type: 'range', indices: [left, right], target };
-
-    const mid = Math.floor((left + right) / 2);
-
-    // Show comparison at mid point
-    yield { type: 'compare', indices: [mid], target, found: false };
-
-    if (arr[mid] === target) {
-      // Found the target
-      yield { type: 'found', indices: [mid], target };
-      return mid;
-    } else if (arr[mid] < target) {
-      left = mid + 1;
-    } else {
-      right = mid - 1;
-    }
-  }
-
-  // Target not found
-  yield { type: 'not-found', target };
-  return -1;
-}
-
-/**
- * Tree Node class for BFS visualization
- */
-class TreeNode {
-  constructor(value) {
-    this.value = value;
-    this.left = null;
-    this.right = null;
-    this.index = null; // For visualization purposes
-  }
-}
-
-/**
- * Creates a binary tree from an array (level-order)
- * @param {number[]} arr - Array to convert to tree
- * @returns {TreeNode|null} - Root of the tree
- */
-function createBinaryTree(arr) {
-  if (!arr || arr.length === 0) return null;
-
-  const root = new TreeNode(arr[0]);
-  root.index = 0;
-  const queue = [root];
-  let i = 1;
-
-  while (queue.length > 0 && i < arr.length) {
-    const node = queue.shift();
-
-    // Add left child
-    if (i < arr.length && arr[i] !== null && arr[i] !== undefined) {
-      node.left = new TreeNode(arr[i]);
-      node.left.index = i;
-      queue.push(node.left);
-    }
-    i++;
-
-    // Add right child
-    if (i < arr.length && arr[i] !== null && arr[i] !== undefined) {
-      node.right = new TreeNode(arr[i]);
-      node.right.index = i;
-      queue.push(node.right);
-    }
-    i++;
-  }
-
-  return root;
-}
-
-/**
- * Breadth-First Search generator for binary tree.
- * Yields objects of the form:
- *   { type: 'visit', index: number, value: number, level: number }
- *   { type: 'queue', indices: number[] } (current queue state)
- *   { type: 'found', index: number, value: number } (if searching for target)
- *   { type: 'complete', traversal: number[] } (final traversal order)
- * @param {number[]} arr - Array representing tree in level-order
- * @param {number} [target] - Optional target value to search for
- * @yields {{type: string, index?: number, indices?: number[], value?: number, level?: number, traversal?: number[], found?: boolean}}
- */
-export function* breadthFirstSearch(arr, target = null) {
-  if (!arr || arr.length === 0) return;
-
-  const root = createBinaryTree(arr);
-  if (!root) return;
-
-  const queue = [{ node: root, level: 0 }];
-  const traversal = [];
-  let found = false;
-
-  while (queue.length > 0) {
-    // Show current queue state
-    const queueIndices = queue.map(item => item.node.index);
-    yield { type: 'queue', indices: queueIndices };
-
-    const { node, level } = queue.shift();
-
-    // Visit current node
-    yield { type: 'visit', index: node.index, value: node.value, level };
-    traversal.push(node.value);
-
-    // Check if this is the target we're searching for
-    if (target !== null && node.value === target) {
-      yield { type: 'found', index: node.index, value: node.value };
-      found = true;
-    }
-
-    // Add children to queue
-    if (node.left) {
-      queue.push({ node: node.left, level: level + 1 });
-    }
-    if (node.right) {
-      queue.push({ node: node.right, level: level + 1 });
-    }
-  }
-
-  // Show final traversal
-  yield { type: 'complete', traversal, found: target !== null ? found : undefined };
-}
-
-/**
- * Depth-First Search generator for binary tree.
- * Yields objects of the form:
- *   { type: 'visit', index: number, value: number, level: number }
- *   { type: 'stack', indices: number[] } (current stack state)
- *   { type: 'found', index: number, value: number } (if searching for target)
- *   { type: 'backtrack', index: number } (when backtracking)
- *   { type: 'complete', traversal: number[] } (final traversal order)
- * @param {number[]} arr - Array representing tree in level-order
- * @param {number} [target] - Optional target value to search for
- * @param {'preorder'|'inorder'|'postorder'} [order='preorder'] - DFS traversal order
- * @yields {{type: string, index?: number, indices?: number[], value?: number, level?: number, traversal?: number[], found?: boolean}}
- */
-export function* depthFirstSearch(arr, target = undefined, order = 'preorder') {
-  if (!arr || arr.length === 0) return;
-
-  const root = createBinaryTree(arr);
-  if (!root) return;
-
-  const traversal = [];
-  let found = false;
-
-  function* dfsHelper(node, level, stack) {
-    if (!node) return;
-
-    // Show current stack state
-    const stackIndices = stack.map(n => n.index);
-    yield { type: 'stack', indices: stackIndices };
-
-    // Add current node to stack for visualization
-    stack.push(node);
-
-    // Preorder: Visit node before children
-    if (order === 'preorder') {
-      yield { type: 'visit', index: node.index, value: node.value, level };
-      traversal.push(node.value);
-
-      // Check if this is the target
-      if (target !== undefined && node.value === target) {
-        yield { type: 'found', index: node.index, value: node.value };
-        found = true;
-        return;
-      }
-    }
-
-    // Recursively visit left subtree
-    if (node.left) {
-      yield* dfsHelper(node.left, level + 1, [...stack]);
-      if (found && target !== undefined) return;
-    }
-
-    // Inorder: Visit node between children
-    if (order === 'inorder') {
-      yield { type: 'visit', index: node.index, value: node.value, level };
-      traversal.push(node.value);
-
-      // Check if this is the target
-      if (target !== undefined && node.value === target) {
-        yield { type: 'found', index: node.index, value: node.value };
-        found = true;
-        return;
-      }
-    }
-
-    // Recursively visit right subtree
-    if (node.right) {
-      yield* dfsHelper(node.right, level + 1, [...stack]);
-      if (found && target !== undefined) return;
-    }
-
-    // Postorder: Visit node after children
-    if (order === 'postorder') {
-      yield { type: 'visit', index: node.index, value: node.value, level };
-      traversal.push(node.value);
-
-      // Check if this is the target
-      if (target !== undefined && node.value === target) {
-        yield { type: 'found', index: node.index, value: node.value };
-        found = true;
-        return;
-      }
-    }
-
-    // Show backtracking
-    yield { type: 'backtrack', index: node.index };
-  }
-
-  // Start DFS
-  yield* dfsHelper(root, 0, []);
-
-  // Show final traversal
-  yield { type: 'complete', traversal, found: target !== undefined ? found : undefined };
-}
